[Refactor] Next.js App 라우터 렌더링 환경 고려 리팩토링
- page.tsx에서 'use client' 삭제
- page-place-page.tsx로 useEffect를 분리

Resolves #121
